Move constant_neq goal
